+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version 0.2.0:
   * Add `Date::add/sub_years` and `Date::add/sub_months`.
   * Add `InvalidDayOfMonth::next/prev_valid()` to get the nearest valid date.
   * Rename `InvalidDate::InvalidDayForMonth` to `InvalidDate::InvalidDayOfMonth` to match the wrapped struct.
+  * Rename `InvalidDayOfYear::day` to `day_of year`.
 
 Version 0.1.1:
   * Fix `Date::add_days` for days in years 100, 200 and 300 of the 400 year cycle.